requirements.txt
python-telegram-bot==20.7
starlette==0.37.2
uvicorn==0.27.1
requests==2.31.0